add comments and refactor verify sig
