+++ api/service/stagedstreamsync/downloader.go
@@ -6,13 +6,9 @@ import (
 	"time"
 
 	"github.com/ethereum/go-ethereum/event"
-	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
 
 	"github.com/harmony-one/harmony/core"
-	"github.com/harmony-one/harmony/core/types"
-	"github.com/harmony-one/harmony/crypto/bls"
-	"github.com/harmony-one/harmony/internal/chain"
 	nodeconfig "github.com/harmony-one/harmony/internal/configs/node"
 	"github.com/harmony-one/harmony/internal/utils"
 	"github.com/harmony-one/harmony/p2p"
@@ -128,7 +124,7 @@ func (d *Downloader) NumPeers() int {
 	return d.syncProtocol.NumStreams()
 }
 
-// IsSyncing returns the current sync status
+// SyncStatus returns the current sync status
 func (d *Downloader) SyncStatus() (bool, uint64, uint64) {
 	syncing, target := d.stagedSyncInstance.status.get()
 	if !syncing {
@@ -251,54 +247,4 @@ func (d *Downloader) loop() {
 			return
 		}
 	}
-}
-
-var emptySigVerifyErr *sigVerifyErr
-
-type sigVerifyErr struct {
-	err error
-}
-
-func (e *sigVerifyErr) Error() string {
-	return fmt.Sprintf("[VerifyHeaderSignature] %v", e.err.Error())
-}
-
-func verifyAndInsertBlocks(bc blockChain, blocks types.Blocks) (int, error) {
-	for i, block := range blocks {
-		if err := verifyAndInsertBlock(bc, block, blocks[i+1:]...); err != nil {
-			return i, err
-		}
-	}
-	return len(blocks), nil
-}
-
-func verifyAndInsertBlock(bc blockChain, block *types.Block, nextBlocks ...*types.Block) error {
-	var (
-		sigBytes bls.SerializedSignature
-		bitmap   []byte
-		err      error
-	)
-	if len(nextBlocks) > 0 {
-		// get commit sig from the next block
-		next := nextBlocks[0]
-		sigBytes = next.Header().LastCommitSignature()
-		bitmap = next.Header().LastCommitBitmap()
-	} else {
-		// get commit sig from current block
-		sigBytes, bitmap, err = chain.ParseCommitSigAndBitmap(block.GetCurrentCommitSig())
-		if err != nil {
-			return errors.Wrap(err, "parse commitSigAndBitmap")
-		}
-	}
-
-	if err := bc.Engine().VerifyHeaderSignature(bc, block.Header(), sigBytes, bitmap); err != nil {
-		return &sigVerifyErr{err}
-	}
-	if err := bc.Engine().VerifyHeader(bc, block.Header(), true); err != nil {
-		return errors.Wrap(err, "[VerifyHeader]")
-	}
-	if _, err := bc.InsertChain(types.Blocks{block}, false); err != nil {
-		return errors.Wrap(err, "[InsertChain]")
-	}
-	return nil
 }
